backend/requirements.txt
flask
flask-cors
pandas
numpy
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
